+++ module/applications/actor/actor-sheet-v2.mjs
@@ -1167,6 +1167,16 @@ export class HeroSystemActorSheetV2 extends HandlebarsApplicationMixin(ActorShee
                     await document.deleteDialog();
                 },
             },
+
+            {
+                name: "Roll (all in one; alpha test)",
+                icon: '<i class="fa-solid fa-fw fa-dice"></i>',
+                condition: () => this.actor.isOwner,
+                callback: async (target) => {
+                    const document = this._getEmbeddedDocument(target);
+                    await document.roll({ event: this.event, token: this.token, allInOne: true });
+                },
+            },
         ];
     }