+++ module/item/item.mjs
@@ -35,7 +35,13 @@ import {
     isManeuverThatDoesReplaceableDamageType,
     isRangedMartialManeuver,
 } from "../utility/damage.mjs";
-import { getRoundedUpDistanceInSystemUnits, getSystemDisplayUnits } from "../utility/units.mjs";
+import {
+    getRoundedUpDistanceInSystemUnits,
+    getSystemDisplayUnits,
+    convertSystemUnitsToMetres,
+    gridUnitsToMeters,
+    currentSceneUsesHexGrid,
+} from "../utility/units.mjs";
 import { HeroRoller } from "../utility/dice.mjs";
 import { HeroSystem6eActorActiveEffects } from "../actor/actor-active-effects.mjs";
 import { getItemDefenseVsAttack } from "../utility/defense.mjs";
@@ -1053,7 +1059,7 @@ export class HeroSystem6eItem extends HeroObjectCacheMixin(Item) {
                 case "TELEKINESIS":
                 case "TRANSFER":
                 case "TRANSFORM":
-                    return collectActionDataBeforeToHitOptions(this, { ...options, event });
+                    return collectActionDataBeforeToHitOptions(this, options);
 
                 case "ABSORPTION":
                 case "DISPEL":
@@ -1084,7 +1090,7 @@ export class HeroSystem6eItem extends HeroObjectCacheMixin(Item) {
                 case "TRIP":
                 default:
                     ui.notifications.warn(`${this.system.XMLID} roll is not fully supported`);
-                    return collectActionDataBeforeToHitOptions(this, event);
+                    return collectActionDataBeforeToHitOptions(this, options);
             }
         } else if (this.baseInfo.behaviors.includes("dice")) {
             switch (this.system.XMLID) {
@@ -6795,6 +6801,56 @@ export class HeroSystem6eItem extends HeroObjectCacheMixin(Item) {
             delete source.system.isHeroic;
             tagObjectForPersistence(source);
         }
+    }
+
+    async placeTemplate(message) {
+        if (!this.aoe) throw Error("Attempted to create template with non-aoe item");
+        return this.placeItemTemplate({ message });
+    }
+
+    async placeItemTemplate(options = {}) {
+        const message = options.message;
+
+        if (!canvas.ready) throw Error("No canvas");
+
+        const areaOfEffect = this.aoeAttackParameters;
+        if (!areaOfEffect) throw Error("No aoeAttackParameters");
+
+        const aoeType = areaOfEffect.type;
+        const aoeValue = areaOfEffect.value;
+        const heroAoeTypeToFoundryAoeTypeConversions = {
+            any: "rect",
+            cone: "cone",
+            line: "ray",
+            radius: "circle",
+            surface: "rect",
+        };
+        const templateType = heroAoeTypeToFoundryAoeTypeConversions[aoeType];
+        const sizeConversionToMeters = convertSystemUnitsToMetres(1, this.actor?.is5e);
+        const hexTemplates = game.settings.get(HEROSYS.module, "HexTemplates");
+        const hexGrid = currentSceneUsesHexGrid();
+
+        // NOTE: If we're using hex templates (i.e. 5e), the target hex is in should count as a distance of 1". This means that to convert to what FoundryVTT expects
+        //       for distance we need to subtract 0.5"/1m from the radius.
+        // NOTE: MeasuredTemplates assume that the distance is in grid units.
+        const distanceInMeters = aoeValue * sizeConversionToMeters - (hexTemplates && hexGrid ? 1 : 0);
+        const distanceInGridUnits = distanceInMeters / gridUnitsToMeters();
+
+        const templateData = {
+            t: templateType,
+            author: game.user.id,
+            distance: distanceInGridUnits,
+            fillColor: game.user.color,
+            flags: {
+                [game.system.id]: {
+                    messageId: message?.id,
+                    purpose: "AoE",
+                    item: this,
+                },
+            },
+        };
+
+        return canvas.templates.createPreview(templateData);
     }
 }